# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoIM 服务器接口文档",
        "title": "GoIM Swagger",
        "termsOfService": "http://yusank.github.io/goim/",
        "contact": {
            "name": "Yusank",
            "url": "https://yusank.space",
            "email": "yusankurban@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/yusank/goim/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/friend/accept-friend": {
            "post": {
                "description": "接受好友请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "接受好友请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ConfirmFriendRequestReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/friend/add-friend": {
            "post": {
                "description": "添加好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "添加好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.AddFriendResult"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/friend/block-friend": {
            "post": {
                "description": "屏蔽好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "屏蔽好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.BaseFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/friend/delete-friend": {
            "post": {
                "description": "删除好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "删除好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.BaseFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/friend/list": {
            "get": {
                "description": "获取好友列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "获取好友列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.QueryFriendListResponse"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/friend/reject-friend": {
            "post": {
                "description": "拒绝好友请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "拒绝好友请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ConfirmFriendRequestReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/friend/unblock-friend": {
            "post": {
                "description": "取消屏蔽好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]好友"
                ],
                "summary": "取消屏蔽好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.BaseFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "err",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/gateway/v1/discovery/discover": {
            "get": {
                "description": "获取推送服务器 agentID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[gateway]discover"
                ],
                "summary": "获取推送服务器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "extra": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "v1.AddFriendRequest": {
            "type": "object",
            "properties": {
                "friend_uid": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "v1.AddFriendResult": {
            "type": "object",
            "properties": {
                "friend_request": {
                    "$ref": "#/definitions/v1.FriendRequest"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "v1.BaseFriendRequest": {
            "type": "object",
            "properties": {
                "friend_uid": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "v1.ConfirmFriendRequestReq": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "integer"
                },
                "friend_request_id": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "v1.Friend": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "friend_avatar": {
                    "type": "string"
                },
                "friend_name": {
                    "type": "string"
                },
                "friend_uid": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "v1.FriendRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "friend_avatar": {
                    "type": "string"
                },
                "friend_name": {
                    "description": "friend info for view",
                    "type": "string"
                },
                "friend_uid": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "v1.QueryFriendListResponse": {
            "type": "object",
            "properties": {
                "friend_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Friend"
                    }
                },
                "response": {
                    "$ref": "#/definitions/response.BaseResponse"
                }
            }
        },
...
    }
}